test: fix `cargo test`
The contents of the `README.md` are included as documentation. The
snippet defined inside of `README.md` are the compiled and run by `cargo
test`.

Running the examples is not a good idea, hence this has been disabled.
On the other hand, compiling the snippets is a good idea. The
`Cargo.toml` had to be update to ensure all the dependencies of the
snippets could be found.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>