Avoid unix-specific functions
To make sure this builds on windows (and other non-unices), we avoid
using the unix-specific OsStr::as_bytes, instead using
OsStr::as_encoded_bytes.

Note that while this is _technically not really valid_, it's only going
to be a problem if the path to the temporary directory contains invalid
utf8 codepoints, which should, hopefully, never happen.