Add `symbolize` feature for online symbolization
This patch adds a `symbolize` feature which performs online
symbolization of profiles using the `backtrace` crate.

This is often much more convenient than having to manually obtain
binaries and symbolize the profiles, and matches the behavior of e.g.
the `pprof-rs` CPU profiler. It will also allow emitting e.g. flamegraph
SVGs directly, which I'll submit in a follow-up PR.

Online symbolization uses a fair amount of memory for symbol caches, but
this is often a worthwhile tradeoff.